# Add term, type, and decl caching to the new codebase format