This adds term, type signature, and type declaration caches to the codebase. (But NOT a branch cache since I think that is very dubious, especially with moving LCA to the database). It improves performance by quite a bit (I'm seeing > 2x speed improvements) and since these caches can be pretty small and still effective, I don't think they affect memory usage much, so this seems like a no brainer.

More compelling is this change the doubles the speed of operations like `find` and `view`, which is important for the codebase server supporting fuzzy find - it has to pull type signatures for all the definitions it's returning, and every millisecond counts there.

It's actually impressive how fast SQLite and our decoding is but it's always going to be faster to just avoid hitting the database and return an object that's already sitting there in memory.

More on the methodology I used, and raw data:

```
Before term and type caching:

stack exec transcripts  91.75s user 32.21s system 93% cpu 2:13.29 total
stack exec transcripts  91.55s user 32.36s system 93% cpu 2:12.85 total
stack exec transcripts -- doc.md  13.96s user 2.28s system 97% cpu 16.670 total
stack exec transcripts -- doc.md  13.74s user 2.16s system 90% cpu 17.645 total
stack exec transcripts -- doc.md  13.75s user 2.14s system 97% cpu 16.346 total

After term and type caching:

stack exec transcripts  80.46s user 30.85s system 92% cpu 2:00.77 total
stack exec transcripts  80.83s user 30.67s system 92% cpu 2:00.86 total
stack exec transcripts -- doc.md  5.94s user 0.93s system 79% cpu 8.616 total
stack exec transcripts -- doc.md  5.98s user 0.91s system 96% cpu 7.127 total
stack exec transcripts -- doc.md  5.96s user 0.95s system 79% cpu 8.695 total
```

Another example I looked at is just the time to do `find` and `view` after initializing a codebase with a bunch of definitions via `builtins.mergeio`. First I measured the time for just `builtins.mergeio`:

````
```ucm:hide
.> builtins.mergeio
```

stack exec transcripts -- scratch.md  3.07s user 0.73s system 95% cpu 3.987 total
````

This time was consistent across runs, let's call it 4s.

Next I added a stanza to call `find` and `view List.map`

````
```ucm
.> find
.> view List.map
```
````

With caching averaging 0.287s (minus 4s for `builtins.mergeio` which typechecks `IOSource`):

```
stack exec transcripts -- scratch.md  3.31s user 0.75s system 95% cpu 4.264 total
stack exec transcripts -- scratch.md  3.34s user 0.76s system 94% cpu 4.325 total
stack exec transcripts -- scratch.md  3.31s user 0.76s system 95% cpu 4.271 total
```

Without caching averaging 0.577s (minus 4s for the `builtins.mergeio`)

```
stack exec transcripts -- scratch.md  3.58s user 0.81s system 96% cpu 4.561 total
stack exec transcripts -- scratch.md  3.58s user 0.82s system 95% cpu 4.593 total
stack exec transcripts -- scratch.md  3.57s user 0.80s system 95% cpu 4.577 total
```