@pchiusano 2021-05-14
Note that CI is failing just due to the current codebase LCA calculation being incorrect. If I add `&& False` it works fine.